Move findSuitableRace implementation to Trackblazer via Campaign hook
